Journey to the Moon
The member states of the UN are planning to send people to the moon. They want them to be from different countries. You will be given a list of pairs of astronaut ids. Each pair is made of astronauts from the same country. Determine how many pairs of astronauts from different countries they can choose from.